Say that 64-bit is required.
Explicitly listing SSE2 (which is included in the x86-64 arch baseline) and
Pentium 4 as minimal requirements makes it sound as if all of them are
supported.  That's not the case as only late Pentium 4 models gained 64 bit.

On the other hand, porting to 32-bit necrocomputing would be a waste of
time (no matter what Steam and Wine think).  Thus, let's correct the README.